remove 1-page form
