--- src/components/ShareExperience/InterviewForm/InterviewExperience/InterviewQa.js
@@ -1,64 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import Textarea from 'react-textarea-autosize';
-
-import AddButton from 'common/button/AddButton';
-
-import styles from './InterviewQa.module.css';
-
-const InterviewQa = ({ question, answer, editQa, removeQa }) => (
-  <div className={styles.container}>
-    <div className={styles.remove__btn}>
-      <AddButton onClick={removeQa} deleteBtn />
-    </div>
-    <div
-      style={{
-        display: 'flex',
-        marginBottom: '14px',
-      }}
-    >
-      <p className={styles.property__title}>Q</p>
-      <input
-        placeholder="面試問題"
-        value={question}
-        onChange={e => editQa('question')(e.target.value)}
-        className={`pM ${styles.input}`}
-        style={{
-          width: '100%',
-        }}
-      />
-    </div>
-    <div
-      style={{
-        display: 'flex',
-        marginBottom: '14px',
-      }}
-    >
-      <p className={styles.property__title}>A</p>
-      <Textarea
-        useCacheForDOMMeasurements
-        value={answer}
-        onChange={e => editQa('answer')(e.target.value)}
-        placeholder="你的回答（選填）"
-        className={styles.textarea}
-        style={{
-          resize: 'none',
-          width: '100%',
-          color: '#333333',
-          fontSize: '1rem',
-          border: 'none',
-          lineHeight: '1.5rem',
-        }}
-      />
-    </div>
-  </div>
-);
-
-InterviewQa.propTypes = {
-  question: PropTypes.string,
-  answer: PropTypes.string,
-  editQa: PropTypes.func,
-  removeQa: PropTypes.func,
-};
-
-export default InterviewQa;
