--- src/components/ShareExperience/InterviewForm/InterviewExperience/InterviewQa.module.css
@@ -1,43 +1,0 @@
-@value main-yellow, main-orange, main-gray from "../../../common/variables.module.css";
-
-.container {
-  position: relative;
-  border: 1px solid transparent;
-  &:hover {
-    border-color: main-yellow;
-
-    > .remove__btn {
-      visibility: visible;
-    }
-  }
-}
-
-.remove__btn {
-  position: absolute;
-  top: 0;
-  right: 0;
-  transform: translate(50%, -50%);
-  visibility: hidden;
-}
-
-.property__title {
-  color: main-orange;
-  font-size: 1.25rem;
-  width: 30px;
-}
-
-.input {
-  color: #969696;
-  border: none;
-
-  &::placeholder {
-    color: main-gray;
-  }
-}
-
-.textarea {
-  &::placeholder {
-    color: main-gray;
-    font-weight: 300;
-  }
-}
