--- src/components/ShareExperience/InterviewForm/InterviewInfo/InterviewInfo.module.css
@@ -1,5 +1,0 @@
-.form {
-  background: white;
-  padding: 49px;
-  box-shadow: 0 0 4px 0 rgba(0, 0, 0, 0.1);
-}
