--- src/components/ShareExperience/InterviewForm/InterviewInfo/InterviewTime.module.css
@@ -1,9 +1,0 @@
-@value warning-red from "../../../common/variables.module.css";
-
-.warning {
-  border: 1px solid warning-red;
-}
-
-.warning__wording {
-  color: warning-red;
-}
