--- src/components/ShareExperience/InterviewForm/InterviewInfo/OverallRating.module.css
@@ -1,27 +1,0 @@
-@value main-yellow from "../../../common/variables.module.css";
-@value warning-red from "../../../common/variables.module.css";
-
-.dialog {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 28px;
-  height: 30px;
-  padding: 4px 7px;
-  background: #fffae1;
-  border: 1px solid main-yellow;
-  border-radius: 3px;
-
-  @media (max-width: 550px) {
-    margin-top: 5px;
-    margin-left: 0;
-  }
-}
-
-.warning {
-  border: 1px solid warning-red;
-}
-
-.warning__wording {
-  color: warning-red;
-}
